fix(build): voice-message fields landed in the wrong model factory
The blanket replace put Message's new fromJson args into
PosProduct.fromJson, which has no such parameters — breaking the release
build. Move them to Message.fromJson and coerce clamp() back to int.

+++ mobile/lib/features/messenger/chat_screen.dart
@@ -81,7 +81,7 @@ class _ChatScreenState extends State<ChatScreen> {
             .read<AppState>()
             .repo
             .sendVoiceMessage(widget.conversation.id, storagePath,
-                secs.clamp(1, 600));
+                secs.clamp(1, 600).toInt());
         if (msg != null && mounted) {
           setState(() => _messages.add(msg));
           _jumpToBottom();
